requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
lxml==5.1.0
pillow==10.2.0
tqdm==4.66.2
selenium==4.18.1
webdriver-manager==4.0.1
pyarrow==15.0.0
openpyxl==3.1.2
tabulate==0.9.0
colorama==0.4.6
python-dateutil==2.8.2